Add functional test (#8)
Test idempotence of clonemachine, i.e. that it creates the same
result when run multiple times.

This is a first functional test. The pattern can be extended to
more tests.

The test checks out a shallow clone of the bitcoin repo. That
means there is only one branch available and for future tests
working on another branch we need to clone again from
another branch. It also means that the initial clone is
much faster.