Disregard some bitcoin files by default (#11)
Some files deviate from upstream bitcoin so much that it doesn't
make sense to try to merge them. Take them from the unit-e repo
instead.

This requires passing in the name of the branch through the
command line option `--unite-branch` in the new executable
`clonemachine.py` (which then calls `fork.py`). If this option is
given a value clonemachine appropriates files specified in its
`appropriated_files` array from this branch.

If the option `--bitcoin-branch` is given as well clonemachine
shows the upstream changes of the appropriated files since the
last merge.

Note that the functional test for this option is not run by the CI
because it requires access to the private `unit-e` repo. You can
execute the test locally. It does fetch both the `unit-e` as well
as the `bitcoin` repo once, though.

This fixes https://github.com/dtr-org/unit-e/issues/146

   Signed-off-by: Julian Fleischer <julian@thirdhash.com>